requirements.txt
lxml
arabic-reshaper
python-bidi